+++ BotService/Services/Interfaces/ITelegramAuthorizationManager.cs
@@ -2,6 +2,6 @@ namespace BotService.Services.Interfaces
 {
     public interface ITelegramAuthorizationManager : IAuthorizationManager
     {
-        bool Authorize(int telegramId);
+        bool Authorize(long telegramId);
     }
 }
